tests/wpt/web-platform-tests/preload/delaying_onload_link_preload_after_discovery.html

<!DOCTYPE html>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
    var t = async_test('Makes sure link preload preloaded resources are delaying onload after discovery');
</script>
<link rel=preload href="resources/dummy.js?pipe=trickle(d5)" as=script>
<link rel=preload href="resources/square.png?pipe=trickle(d5)" as=image>
<body>
<script>
    window.addEventListener("load", t.step_func(function() {
        assert_equals(performance.getEntriesByType("resource").length, 4);
        t.done();
    }));
    var script = document.createElement("script");
    script.src = "resources/dummy.js?pipe=trickle(d5)";
    document.body.appendChild(script);
    var img = new Image();
    img.src = "resources/square.png?pipe=trickle(d5)";
</script>
